Feat: Initialize backend with main application and routing setup
