+++ backend/main.go
@@ -1,0 +1,52 @@
+package main
+
+import (
+	"fmt"
+	"log"
+
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+	docs "github.com/prathamrao021/HelperHub/docs"
+	"github.com/prathamrao021/HelperHub/models"
+	"github.com/prathamrao021/HelperHub/routes"
+	swaggerFiles "github.com/swaggo/files"
+	ginSwagger "github.com/swaggo/gin-swagger"
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
+)
+
+func initDB() *gorm.DB {
+	var err error
+	dsn := "host=localhost user=postgres password=admin dbname=User port=5432 sslmode=prefer TimeZone=Asia/Shanghai"
+	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
+	if err != nil {
+		log.Fatal("Failed to connect to database:", err)
+	}
+
+	db.AutoMigrate(&models.User{})
+	fmt.Println("Database connection successfully opened")
+	return db
+}
+
+func main() {
+	router := gin.Default()
+
+	router.Use(cors.New(cors.Config{
+		AllowOrigins:     []string{"*"},
+		AllowMethods:     []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
+		AllowHeaders:     []string{"Origin", "Content-Type", "Authorization"},
+		ExposeHeaders:    []string{"Content-Length"},
+		AllowCredentials: true,
+	}))
+
+	// db := initDB()
+	var db *gorm.DB
+
+	routes.SetupRoutes(router, db)
+
+	docs.SwaggerInfo.BasePath = "/"
+	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
+
+	log.Print("Starting server on port 8080")
+	router.Run(":8080")
+}
