+++ backend/routes/routes.go
@@ -1,38 +1,15 @@
-package main
+package routes
 
 import (
-	"fmt"
-	"log"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 	"github.com/prathamrao021/HelperHub/models"
 	"golang.org/x/crypto/bcrypt"
-	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
 
-var db *gorm.DB
-
-func initDB() {
-	var err error
-	dsn := "host=localhost user=postgres password=admin dbname=User port=5432 sslmode=prefer TimeZone=Asia/Shanghai"
-	db, err = gorm.Open(postgres.Open(dsn), &gorm.Config{})
-	if err != nil {
-		log.Fatal("Failed to connect to database:", err)
-	}
-
-	db.AutoMigrate(&models.User{})
-	fmt.Println("Database connection successfully opened")
-}
-
-func Caller(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{
-		"message": "pong",
-	})
-}
-
-func createUser(c *gin.Context) {
+func createUser(c *gin.Context, db *gorm.DB) {
 	var user models.User
 	if err := c.ShouldBindJSON(&user); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
@@ -55,7 +32,7 @@ func createUser(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"message": "User data created successfully"})
 }
 
-func deleteUser(c *gin.Context) {
+func deleteUser(c *gin.Context, db *gorm.DB) {
 	username := c.Param("username")
 
 	if err := db.Where("username = ?", username).Delete(&models.User{}).Error; err != nil {
@@ -66,11 +43,10 @@ func deleteUser(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"message": "User deleted successfully"})
 }
 
-func main() {
-	initDB()
-	r := gin.Default()
-	r.GET("/ping", Caller)
-	r.POST("/users/create", createUser)
-	r.POST("/users/delete/:username", deleteUser)
-	r.Run() // listen and serve on 0.0.0.0:8080 (for windows "localhost:8080")
+func SetupRoutes(router *gin.Engine, db *gorm.DB) {
+
+	// Routes for user management
+	userRouter := router.Group("/users")
+	userRouter.POST("/create", func(c *gin.Context) { createUser(c, db) })
+	userRouter.POST("/delete/:username", func(c *gin.Context) { deleteUser(c, db) })
 }
